+++ CHANGELOG.md
@@ -3,7 +3,8 @@ All Notable changes to the `chief` application template will be documented in th
 
 ## unreleased
 - Added: A Page can now set a fixed base url segment (e.g. /news). This is set on the Page model via a `baseUrlSegment` property. This accepts a single string or an array of localized segments where the key should be the locale itself.
-- Changed: `ActsAsMenuItem::mediaUrl()`  and `Page::mediaUrl()` are deprecated. Use `url()` method instead.
+- Added: pages.home route is not configurable in the chief config file.
+- Changed: `ActsAsMenuItem::menuUrl()`  and `Page::menuUrl()` are deprecated. Use `url()` method instead.
 
 ## 0.2.7 - 2019-03-21
 - Added: Publish filter added to pagemanager.